agent/requirements.txt
flask
flask-cors
python-dotenv
requests
gTTS
google-genai